Re-organized some errors to avoid circular imports. The structure of this package is going to change quite a bit here
